flask_api/requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
imutils==0.5.3
numpy==1.18.1
pandas==1.0.0
Pillow==7.0.0
image==1.5.28
opencv-python-headless==4.2.0.32
pytesseract==0.3.2
absl-py==0.7.0
astor==0.7.1
editdistance==0.5.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
Markdown==3.0.1
numpy==1.16.1
opencv-python==4.0.0.21
protobuf==3.6.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
